Failing but mostly magic movegen
